[AR-05][AR-07][AR-08] Consolidate schemas, break MCP cycle, and fix ADR hygiene
