+++ docs/architecture/ADRs.md
@@ -36,6 +36,7 @@ This document indexes all architectural decisions for the Holiday Peak Hub accel
 | [ADR-028](adrs/adr-028-memory-namespace-isolation-contract.md) | Memory Namespace Isolation Contract | Accepted | 2026-03 |
 | [ADR-029](adrs/adr-029-dam-image-analysis-enrichment-pipeline-boundary.md) | DAM Image Analysis as an Enrichment Pipeline within Truth Boundaries | Accepted | 2026-03 |
 | [ADR-030](adrs/adr-030-search-enrichment-bounded-context.md) | Search Enrichment as an Isolated Bounded Context | Accepted | 2026-03 |
+| [ADR-031](adrs/adr-031-mcp-internal-communication-policy.md) | MCP Internal Communication Policy Addendum | Accepted | 2026-03 |
 
 ## How to Use ADRs
 
@@ -99,7 +100,7 @@ Each ADR follows a standard template:
 
 ### Governance
 - Git branch naming convention ([ADR-022](adrs/adr-022-branch-naming-convention.md))
-- MCP internal communication policy and rollout gates ([ADR-029](adrs/adr-029-mcp-internal-communication-policy.md))
+- MCP internal communication policy and rollout gates ([ADR-031](adrs/adr-031-mcp-internal-communication-policy.md))
 
 ### Enterprise Integration
 - Enterprise resilience patterns (Circuit Breaker, Bulkhead, Rate Limiter) ([ADR-023](adrs/adr-023-enterprise-resilience-patterns.md))
