+++ lib/src/holiday_peak_lib/agents/fastapi_mcp.py
@@ -1,17 +1,5 @@
-"""Thin wrapper to expose MCP endpoints over FastAPI."""
+"""Backward-compatible import location for FastAPIMCPServer."""
 
-from fastapi import APIRouter, FastAPI
+from holiday_peak_lib.mcp.server import FastAPIMCPServer
 
-
-class FastAPIMCPServer:
-    """Registers MCP routes on a FastAPI app."""
-
-    def __init__(self, app: FastAPI) -> None:
-        self.app = app
-        self.router = APIRouter()
-
-    def add_tool(self, path: str, handler) -> None:
-        self.router.post(path)(handler)
-
-    def mount(self) -> None:
-        self.app.include_router(self.router, prefix="/mcp")
+__all__ = ["FastAPIMCPServer"]
