+++ lib/src/holiday_peak_lib/mcp/__init__.py
@@ -5,9 +5,11 @@ from holiday_peak_lib.mcp.ai_search_indexing import (
     build_ai_search_indexing_client_from_env,
     register_ai_search_indexing_tools,
 )
+from holiday_peak_lib.mcp.server import FastAPIMCPServer
 
 __all__ = [
     "AISearchIndexingClient",
     "build_ai_search_indexing_client_from_env",
     "register_ai_search_indexing_tools",
+    "FastAPIMCPServer",
 ]
